[feat]: 에디터 컴포넌트화 (#62)
