+++ src/pages/QnaDetailsPage.tsx
@@ -1,13 +1,17 @@
+import { useState } from 'react'
 import DetailsHeader from '@/components/details/DetailsHeader'
 import DetailsContents from '@/components/details/DetailsContents'
 import DetailsWriter from '@/components/details/DetailsWriter'
 import DetailsAnswerList from '@/components/details/DetailsAnswerList'
 import AiAnswerSection from '@/components/details/DetailsChatbot'
+import ProfileImage from '@/assets/images/svg/ProfileThumb.svg'
 import { useQuery } from '@tanstack/react-query'
 import { FetchQnaDetails } from '@/hooks/FetchQnaDetails'
 import { useParams } from 'react-router'
 export default function QnaDetailsPage() {
   const { id } = useParams<{ id: string }>()
+  const [showAnswerForm, setShowAnswerForm] = useState(false)
+
   const { data, isLoading, isError } = useQuery({
     queryKey: ['qnaDetails', id],
     queryFn: () => FetchQnaDetails(Number(id)),
@@ -27,7 +31,32 @@ export default function QnaDetailsPage() {
       />
       <DetailsContents content={data.content} />
       <AiAnswerSection />
-      <DetailsWriter />
+
+      {!showAnswerForm && (
+        <div className="mt-[52px]">
+          <div className="flex items-center justify-between rounded-[20px] border border-[#E5E7EB] bg-white px-[38px] py-[24px]">
+            <div className="flex items-center gap-[12px]">
+              <img src={ProfileImage} alt="프로필 이미지" className="h-[48px] w-[48px]" />
+              <div>
+                <span className="text-[12px] text-[#6201E0]">오즈오즈 님,</span>
+                <p className="text-[18px] font-semibold text-[#222]">
+                  정보를 공유해 주세요.
+                </p>
+              </div>
+            </div>
+            <button
+              type="button"
+              onClick={() => setShowAnswerForm(true)}
+              className="rounded-[8px] bg-[#7C3AED] px-[24px] py-[12px] text-[14px] font-semibold text-white transition-colors hover:bg-[#6D28D9]"
+            >
+              답변하기
+            </button>
+          </div>
+        </div>
+      )}
+
+      {showAnswerForm && <DetailsWriter />}
+
       <DetailsAnswerList answers={data.answers} />
     </div>
   )
